sosta: A package for the analysis of morphologically-relevant structures in spatial omics data

`sosta` (Spatial Omics STructure Analysis) is a package for analyzing spatial omics data to explore tissue organization at the anatomical structure level. It reconstructs morphologically relevant structures based on molecular features or cell types. It further calculates a range of structural and shape metrics to quantitatively describe tissue architecture. The package is designed to integrate with other packages for the analysis of spatial (omics) data.

See also

Author

Maintainer: Samuel Gunz samuel.gunz@uzh.ch (ORCID)

Authors:

Mark D. Robinson mark.robinson@uzh.ch [funder]