Calculate a set of shape metrics of a set of polygons

Calculate a set of shape metrics of a set of polygons

Usage

totalShapeMetrics(sfInput)

Arguments

sfInput

MULTIPOLYGON of class sf

Value

matrix; matrix of shape metrics

Details

Calculate a set of shape metrics of a set of polygons. The function calculates all metrics that are implemented in the function shapeMetrics()

Examples

spe <- imcdatasets::Damond_2019_Pancreas("spe", full_dataset = FALSE)
#> see ?imcdatasets and browseVignettes('imcdatasets') for documentation
#> loading from cache
islet_poly <- reconstructShapeDensityImage(spe,
    marks = "cell_category",
    image_col = "image_name", image_id = "E04", mark_select = "islet", dim = 500
)
totalShapeMetrics(islet_poly)
#>               islet_poly1  islet_poly2  islet_poly3 islet_poly4
#> Area         7325.9386763 6625.0693932 2.583162e+04 501.7365063
#> Compactness     1.8704159    1.8201399 1.992508e+00   1.7395578
#> Eccentricity    0.6132393    0.6132393 6.132393e-01   0.6132393
#> Circularity     0.8728402    0.8770245 8.495254e-01   0.9083714
#> Solidity        0.9794454    0.9657773 9.539998e-01   0.9492151
#> Curl            0.2189454    0.1583197 6.060057e-01  -2.1911744
#> fibreLength   162.3574421  150.6629265 3.218575e+02  39.7377288
#> fibreWidth     45.1222844   43.9727911 8.025794e+01  12.6261999