Calculate mean shape metrics of a set of polygons

Calculate mean shape metrics of a set of polygons

Usage

meanShapeMetrics(totalShapeMetricMatrix)

Arguments

totalShapeMetricMatrix

matrix of shape metrics

Value

matrix; matrix of mean shape metrics

Examples

spe <- imcdatasets::Damond_2019_Pancreas("spe", full_dataset = FALSE)
#> see ?imcdatasets and browseVignettes('imcdatasets') for documentation
#> loading from cache
islet_poly <- reconstructShapeDensityImage(spe,
    marks = "cell_category",
    image_col = "image_name", image_id = "E04", mark_select = "islet", dim = 500
)
shape_metrics <- totalShapeMetrics(islet_poly)
meanShapeMetrics(shape_metrics)
#>                          [,1]
#> Area             1.007109e+04
#> Compactness      1.855655e+00
#> Eccentricity     6.812548e-01
#> Circularity      8.769404e-01
#> Solidity         9.621094e-01
#> Curl             2.542493e-01
#> fibreLength      1.686539e+02
#> fibreWidth       4.549480e+01
#> numberStructures 4.000000e+00
#> totalArea        4.028437e+04